親コンポーネントからデータを渡すように修正
